requirements.txt
# Django
Django==4.1
django-cors-headers==3.13.0
django-database-url==1.0.3
whitenoise==6.2.0
django-filter==22.1
django-phonenumber-field[phonenumbers]==6.3.0
djangorestframework==3.13.1

# Postgres
psycopg2-binary==2.9.3

# Deployment
uvicorn==0.18.2
gunicorn==20.1.0

# tzdata==2022.1
